Extract `classifieds::utils` module
